+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- `markdown.marp.pdf.outlines` option to generate outlines in exported PDF ([#381](https://github.com/marp-team/marp-vscode/issues/381), [#384](https://github.com/marp-team/marp-vscode/pull/384))
+
 ### Changed
 
 - Upgrade Marp Core to [v3.3.3](https://github.com/marp-team/marp-core/releases/tag/v3.3.3) ([#378](https://github.com/marp-team/marp-vscode/pull/378))
-- Upgrade Marp CLI to [v2.1.4](https://github.com/marp-team/marp-cli/releases/tag/v2.1.4) ([#380](https://github.com/marp-team/marp-vscode/pull/380))
+- Upgrade Marp CLI to [v2.2.0](https://github.com/marp-team/marp-cli/releases/tag/v2.2.0) ([#384](https://github.com/marp-team/marp-vscode/pull/384))
 
 ## v2.2.1 - 2022-08-12